Refactored single client into RPC client base class
