+++ Assets/NPCEngine/Scripts/RPC/RPCBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b68b40d3c15a422478d329d9428c8a14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
