+++ Assets/NPCEngine/Scripts/RPC/RPCTypes.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6a4e2dad63d287d47b74a8d98c4cc196
+guid: 3f755cf2320f82c4496d8cdc90d47960
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
